Change TimestampStyle enum underlying type to byte
The `TimestampStyle` enum in the `NetCord` namespace has been
modified. The underlying type of the enum has been changed from
`short` to `byte`. Additionally, the values assigned to each enum
member have been updated to cast from `byte` instead of `short`.
This change reduces the memory footprint of the enum values from
2 bytes to 1 byte per value.

+++ NetCord/TimestampStyle.cs
@@ -1,12 +1,12 @@
 ﻿namespace NetCord;
 
-public enum TimestampStyle : short
+public enum TimestampStyle : byte
 {
-    ShortTime = (short)'t',
-    LongTime = (short)'T',
-    ShortDate = (short)'d',
-    LongDate = (short)'D',
-    ShortDateTime = (short)'f',
-    LongDateTime = (short)'F',
-    RelativeTime = (short)'R',
+    ShortTime = (byte)'t',
+    LongTime = (byte)'T',
+    ShortDate = (byte)'d',
+    LongDate = (byte)'D',
+    ShortDateTime = (byte)'f',
+    LongDateTime = (byte)'F',
+    RelativeTime = (byte)'R',
 }
